+++ CHANGELOG.md
@@ -8,8 +8,11 @@ This project aims to adhere to [Semantic Versioning](http://semver.org/).
  - Moved configuration validation into a static method on ConfigContext.
  - Removed MantaCryptoException.
  - Manta specific content-types are now contained in `MantaContentTypes`.
+ - MantaClient.close() no longer throws Exception.
 ### Added
  - Added version information to user agent and debug output. 
+ - Added simple CLI to use for debugging.
+ - Added configuration dump system property: manta.configDump
 
 ## [2.7.2] - ?
 ### Changed